requirements.txt
pillow == 9.*
requests == 2.*
selenium == 4.*
